Changes: allow undef data type, added date field
Modified enumerate.py to allow an undefined data type. This is
particularly useful for custom parsers. Also added a date field
for enumerate to allow parsers to extract the date information
from a particular file.

Finally, added a warning about the deprecation of enumerate_all.
This function (and its corresponding arguments and options
elsewhere) will be removed in the near future.